refactor(car): CarService를 car 패키지 안으로 이동
